+++ src/test/groovy/ru/vyarus/dropwizard/guice/config/debug/AvailableInstallersPrintTest.groovy
@@ -29,7 +29,6 @@ class AvailableInstallersPrintTest extends Specification {
         @Override
         void initialize(Bootstrap<Configuration> bootstrap) {
             bootstrap.addBundle(GuiceBundle.builder()
-                    .useWebInstallers()
                     .printAvailableInstallers()
                     .build())
         }